refactor(remote): Add package remote,flow and example
