+++ zio-flow/shared/src/main/scala/zio/flow/examples/Example.scala
@@ -1,6 +1,9 @@
-package zio.flow
+package zio.flow.examples
 
-import zio.schema.{ DeriveSchema, Schema }
+import zio.flow._
+import zio.flow.remote.Remote
+import zio.flow.zFlow.ZFlow
+import zio.schema.{DeriveSchema, Schema}
 
 object Example {
   // Remote[A] => Remote[(B, A)]
