+++ zio-flow/shared/src/main/scala/zio/flow/examples/HelloWorld.scala
@@ -1,9 +1,9 @@
-package zio.flow
+package zio.flow.examples
+
+import zio.console.{Console, getStrLn, putStrLn}
+import zio.{App, ExitCode, URIO, ZIO}
 
 import java.io.IOException
-
-import zio.console._
-import zio.{ App, ExitCode, URIO, ZIO }
 
 object HelloWorld extends App {
 
@@ -17,8 +17,8 @@ object HelloWorld extends App {
 
   val myAppLogic: ZIO[Console, IOException, Unit] =
     for {
-      _    <- putStrLn("Hello! What is your name?")
+      _ <- putStrLn("Hello! What is your name?")
       name <- getStrLn
-      _    <- putStrLn(s"Hello, $name, welcome to ZIO!")
+      _ <- putStrLn(s"Hello, $name, welcome to ZIO!")
     } yield ()
 }
