+++ zio-flow/shared/src/main/scala/zio/flow/remote/RemoteDurationSyntax.scala
@@ -1,7 +1,7 @@
-package zio.flow
+package zio.flow.remote
 
 import java.time.Duration
-import java.time.temporal.{ Temporal, TemporalAmount }
+import java.time.temporal.{Temporal, TemporalAmount}
 
 class RemoteDurationSyntax(val self: Remote[Duration]) extends AnyVal {
