+++ zio-flow/shared/src/main/scala/zio/flow/remote/RemoteExecutingFlowSyntax.scala
@@ -1,4 +1,7 @@
-package zio.flow
+package zio.flow.remote
+
+import zio.flow.zFlow.ZFlow
+import zio.flow.{ActivityError, ExecutingFlow, FlowId}
 
 class RemoteExecutingFlowSyntax[A](self: Remote[A]) {
