+++ zio-flow/shared/src/main/scala/zio/flow/remote/RemoteRelationalSyntax.scala
@@ -1,4 +1,4 @@
-package zio.flow
+package zio.flow.remote
 
 class RemoteRelationalSyntax[A](self: Remote[A]) {
