+++ zio-flow/shared/src/main/scala/zio/flow/remote/RemoteStringSyntax.scala
@@ -1,4 +1,4 @@
-package zio.flow
+package zio.flow.remote
 
 class RemoteStringSyntax(self: Remote[String]) {
   def length: Remote[Int] = Remote.Length(self)
