+++ zio-flow/shared/src/main/scala/zio/flow/remote/RemoteVariableSyntax.scala
@@ -1,4 +1,7 @@
-package zio.flow
+package zio.flow.remote
+
+import zio.flow.Variable
+import zio.flow.zFlow.ZFlow
 
 class RemoteVariableSyntax[A](val self: Remote[Variable[A]]) extends AnyVal {
   def get: ZFlow[Any, Nothing, A] = self.modify((a: Remote[A]) => (a, a))
