+++ zio-flow/shared/src/main/scala/zio/flow/remote/numerics.scala
@@ -1,4 +1,4 @@
-package zio.flow
+package zio.flow.remote
 
 import zio.schema.Schema
