+++ zio-flow/shared/src/main/scala/zio/flow/zFlow/ZFlow.scala
@@ -1,9 +1,11 @@
-package zio.flow
+package zio.flow.zFlow
 
-import java.time.{ Duration, Instant }
-
-import zio.flow.ZFlow.Die
+import zio.flow.remote.{Remote, RemoteEitherSyntax}
+import zio.flow.zFlow.ZFlow._
+import zio.flow.{Activity, ActivityError, ExecutingFlow, Variable}
 import zio.schema.Schema
+
+import java.time.{Duration, Instant}
 
 // ZFlow - models a workflow
 //  - terminate, either error or value
