+++ zio-flow/shared/src/main/scala/zio/flow/zFlow/ZFlowTransaction.scala
@@ -1,4 +1,6 @@
-package zio.flow
+package zio.flow.zFlow
+
+import zio.flow.remote.Remote
 
 sealed trait ZFlowTransaction {
