+++ zio-flow/shared/src/test/scala/zio/flow/ZFlowExecutorSpec.scala
@@ -2,16 +2,17 @@ package zio.flow
 
 import java.net.URI
 import java.time.Instant
-
-import zio.flow.ZFlowExecutor.InMemory.{ CompileStatus, State, TState }
+import zio.flow.zFlow.ZFlowExecutor.InMemory.{CompileStatus, State, TState}
+import zio.flow.remote.Remote
 import zio.flow.utils.ZFlowAssertionSyntax.InMemoryZFlowAssertion
 import zio.flow.utils.ZFlowAssertionSyntax.Mocks.mockInMemoryTestClock
+import zio.flow.zFlow.ZFlow
 import zio.schema.DeriveSchema.gen
 import zio.schema.Schema
-import zio.test.Assertion.{ dies, equalTo, fails, hasMessage }
+import zio.test.Assertion.{dies, equalTo, fails, hasMessage}
 import zio.test.TestAspect.ignore
 import zio.test._
-import zio.{ Promise, Ref }
+import zio.{Promise, Ref}
 
 object ZFlowExecutorSpec extends DefaultRunnableSpec {
