+++ zio-flow/shared/src/test/scala/zio/flow/utils/RemoteAssertionSyntax.scala
@@ -1,8 +1,8 @@
 package zio.flow.utils
 
-import zio.flow.Remote
-import zio.test.Assertion.{ equalTo, isRight }
-import zio.test.{ TestResult, assert }
+import zio.flow.remote.Remote
+import zio.test.Assertion.{equalTo, isRight}
+import zio.test.{TestResult, assert}
 
 object RemoteAssertionSyntax {
