+++ zio-flow/shared/src/test/scala/zio/flow/utils/ZFlowAssertionSyntax.scala
@@ -1,15 +1,15 @@
 package zio.flow.utils
 
 import java.net.URI
-
 import zio.clock.Clock
 import zio.console.Console
-import zio.flow.ZFlowExecutor.InMemory
+import zio.flow.zFlow.ZFlowExecutor.InMemory
 import zio.flow.utils.MocksForGCExample.mockInMemoryForGCExample
-import zio.flow.{ Activity, ActivityError, Operation, OperationExecutor, ZFlow }
+import zio.flow.zFlow.ZFlow
+import zio.flow.{Activity, ActivityError, Operation, OperationExecutor}
 import zio.schema.DeriveSchema.gen
 import zio.schema.Schema
-import zio.{ Has, Ref, ZIO, console }
+import zio.{Has, Ref, ZIO, console}
 
 object ZFlowAssertionSyntax {
